requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djoser==2.2.2
django-storages==1.13.2
boto3==1.28.0
python-dotenv==0.19.1
Pillow==9.0.0
